Appdemercados v2.0.7 - Sistema de votação em comentários e histórico completo
- Adicionado sistema de likes/dislikes em comentários do DAO
- Implementado impacto de votos na reputação dos usuários (+0.1 por like, -0.1 por dislike)
- Criado modal de histórico completo com 3 abas (Transações, Reputação, Atividades)
- Adicionado endpoint para listar validações de reputação recebidas
- Modificado Service Worker para nunca cachear HTML/JS
- Alterados schemas de reputação de int para float para suportar decimais
- Adicionado filtro de sugestões por usuário no endpoint DAO
- Implementado exibição de sugestões do usuário na carteira
- Adicionado botão colapsável "Como ganhar tokens?" na carteira
- Exibição de reputação em números inteiros no frontend (mantendo decimais no backend)
- Criada tabela votos_comentarios no banco de dados

+++ app/models/schemas_crypto.py
@@ -57,7 +57,7 @@ class SaldoResponse(BaseModel):
     total_gasto: float
     ultima_atualizacao: datetime
     # Reputação
-    reputacao: int = 100
+    reputacao: float = 100
     total_validacoes_feitas: int = 0
     total_validacoes_recebidas: int = 0
     validacoes_positivas: int = 0
